Delete old mixins.
