--- packages/addon-mixins/.eslintrc
@@ -1,7 +1,0 @@
-{
-  "rules": {
-    "no-magic-numbers": "off",
-    "sort-keys": "off",
-    "babel/no-invalid-this": "off"
-  }
-}
