--- packages/addon-mixins/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@aesthetic/addon-mixins",
-  "version": "0.2.11",
-  "description": "CSS-in-JS mixins for the Aesthetic design system.",
-  "keywords": [
-    "aesthetic",
-    "addon",
-    "mixins",
-    "css",
-    "design",
-    "system"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:aesthetic-suite/framework.git",
-    "directory": "packages/addon-mixins"
-  },
-  "author": "Miles Johnson",
-  "license": "MIT",
-  "main": "./lib/index.js",
-  "module": "./esm/index.js",
-  "types": "./dts/index.d.ts",
-  "files": [
-    "dts/**/*.d.ts",
-    "esm/**/*.{js,map}",
-    "lib/**/*.{js,map}",
-    "src/**/*.{ts,tsx,json}"
-  ],
-  "sideEffects": false,
-  "publishConfig": {
-    "access": "public"
-  },
-  "peerDependencies": {
-    "@aesthetic/sss": "*",
-    "@aesthetic/system": "*"
-  },
-  "dependencies": {
-    "@aesthetic/utils": "^0.7.0"
-  },
-  "funding": {
-    "type": "ko-fi",
-    "url": "https://ko-fi.com/milesjohnson"
-  },
-  "packemon": {
-    "platform": "browser"
-  }
-}
