+++ packages/system/src/Design.ts
@@ -1,11 +1,9 @@
 import { deepMerge } from '@aesthetic/utils';
 import { DEPTHS } from './constants';
 import Theme from './Theme';
-import { DeepPartial, DesignTokens, MixinTemplateMap, ThemeOptions, ThemeTokens } from './types';
+import { DeepPartial, DesignTokens, ThemeOptions, ThemeTokens } from './types';
 
 export default class Design<T extends object> {
-  readonly mixins?: MixinTemplateMap<T>;
-
   readonly name: string;
 
   readonly rootLineHeight: number;
@@ -16,9 +14,8 @@ export default class Design<T extends object> {
 
   readonly tokens: DesignTokens;
 
-  constructor(name: string, tokens: Omit<DesignTokens, 'depth'>, mixins?: MixinTemplateMap<T>) {
+  constructor(name: string, tokens: Omit<DesignTokens, 'depth'>) {
     this.name = name;
-    this.mixins = mixins;
     this.tokens = { ...tokens, depth: DEPTHS };
     this.rootLineHeight = tokens.typography.rootLineHeight;
     this.rootTextSize = Number.parseFloat(tokens.typography.rootTextSize);
@@ -37,7 +34,7 @@ export default class Design<T extends object> {
   /**
    * Extend and instantiate a new design instance with customized design tokens.
    */
-  extend(name: string, tokens: DeepPartial<DesignTokens>, mixins?: MixinTemplateMap<T>): Design<T> {
-    return new Design(name, deepMerge(this.tokens, tokens), { ...this.mixins, ...mixins });
+  extend(name: string, tokens: DeepPartial<DesignTokens>): Design<T> {
+    return new Design(name, deepMerge(this.tokens, tokens));
   }
 }
