+++ packages/system/src/Theme.ts
@@ -3,11 +3,10 @@
 import { ColorScheme, ContrastLevel, Unit, VariablesMap } from '@aesthetic/types';
 import { deepMerge, hyphenate, isObject, objectLoop } from '@aesthetic/utils';
 import Design from './Design';
+import { MIXIN_MAP } from './mixins';
 import {
   DeepPartial,
-  MixinTemplate,
-  MixinTemplateMap,
-  MixinUtil,
+  MixinType,
   ThemeOptions,
   ThemeTokens,
   Tokens,
@@ -16,50 +15,31 @@ import {
 } from './types';
 
 type AnyObject = Record<string, any>;
-type MixinTemplateCache<T extends object> = Record<string, Set<MixinTemplate<T>>>;
 
 export default class Theme<T extends object> implements Utilities<T> {
   name: string = '';
 
   readonly contrast: ContrastLevel;
 
-  readonly mixin: MixinUtil<T>;
-
   readonly scheme: ColorScheme;
 
   readonly tokens: Tokens;
-
-  protected mixins: Record<string, MixinTemplate<T>> = {};
-
-  protected templates: MixinTemplateCache<T> = {};
 
   private cssVariables?: VariablesMap;
 
   private design: Design<T>;
 
-  constructor(
-    options: ThemeOptions,
-    tokens: ThemeTokens,
-    design: Design<T>,
-    parentTemplates: MixinTemplateCache<T> = {},
-  ) {
+  constructor(options: ThemeOptions, tokens: ThemeTokens, design: Design<T>) {
     this.contrast = options.contrast;
     this.scheme = options.scheme;
     this.design = design;
     this.tokens = { ...design.tokens, ...tokens };
-    this.templates = parentTemplates;
 
     // Bind instead of using anonymous functions since we need
     // to pass these around and define properties on them!
-    this.mixin = this.mixinBase.bind(this);
+    this.mixin = this.mixin.bind(this);
     this.unit = this.unit.bind(this);
     this.var = this.var.bind(this);
-
-    // If the parent design has defined mixins,
-    // register them as special built-ins.
-    if (design.mixins) {
-      this.registerBuiltIns(design.mixins);
-    }
   }
 
   /**
@@ -74,32 +54,7 @@ export default class Theme<T extends object> implements Utilities<T> {
       },
       deepMerge(this.tokens, tokens),
       this.design,
-      { ...this.templates },
     );
-  }
-
-  /**
-   * Extend a registered mixin with additional CSS properties.
-   */
-  extendMixin(name: string, template: MixinTemplate<T>): this {
-    this.templates[name] = (this.templates[name] || new Set()).add(template);
-
-    return this;
-  }
-
-  /**
-   * Register a mixin to provide reusable CSS properties.
-   */
-  registerMixin(name: string, template: MixinTemplate<T>): this {
-    if (__DEV__) {
-      if (this.mixins[name]) {
-        throw new Error(`A mixin already exists for "${name}". Cannot overwrite.`);
-      }
-    }
-
-    this.mixins[name] = template;
-
-    return this;
   }
 
   /**
@@ -134,38 +89,18 @@ export default class Theme<T extends object> implements Utilities<T> {
    * Return merged CSS properties from the defined mixin, all template overrides,
    * and the provided additional CSS properties.
    */
-  mixinBase(name: string, maybeOptions?: T | object, maybeRule?: T): T {
-    let options: object = {};
-    let rule: T | undefined;
-
-    if (maybeOptions && maybeRule) {
-      options = maybeOptions;
-      rule = maybeRule;
-    } else if (maybeOptions) {
-      rule = maybeOptions as T;
-    }
-
+  mixin(name: MixinType, rule?: T): T {
     const properties: object = {};
-    const mixin = this.mixins[name];
+    const mixin = MIXIN_MAP[name];
 
     if (mixin) {
-      Object.assign(properties, mixin.call(this, options));
+      Object.assign(properties, mixin.call(this));
     } else if (__DEV__) {
-      // Log instead of error since mixins are dynamic
-      // eslint-disable-next-line no-console
-      console.warn(`Unknown mixin "${name}".`);
-    }
-
-    const templates = this.templates[name];
-
-    if (templates) {
-      templates.forEach((template) => {
-        Object.assign(properties, template.call(this, options));
-      });
+      throw new Error(`Unknown mixin "${name}".`);
     }
 
     if (rule) {
-      Object.assign(properties, rule);
+      return deepMerge(properties, rule);
     }
 
     return properties as T;
@@ -190,22 +125,5 @@ export default class Theme<T extends object> implements Utilities<T> {
    */
   var(name: VariableName, ...fallbacks: (number | string)[]): string {
     return `var(${[`--${name}`, ...fallbacks].join(', ')})`;
-  }
-
-  /**
-   * Register the built-in mixin's as properties on the mixin method for easy access.
-   */
-  protected registerBuiltIns(mixins: MixinTemplateMap<T>) {
-    objectLoop(mixins, (template, name) => {
-      const type = hyphenate(name);
-
-      // Register the mixin
-      this.registerMixin(type, template);
-
-      // Provide a utility function
-      Object.defineProperty(this.mixin, name, {
-        value: (options?: object, rule?: T) => this.mixin(type, options!, rule!),
-      });
-    });
   }
 }
