+++ packages/system/src/index.ts
@@ -4,10 +4,11 @@
  */
 
 import Design from './Design';
+import * as mixin from './mixins';
 import Theme from './Theme';
 import ThemeRegistry from './ThemeRegistry';
 
 export * from './constants';
 export * from './types';
 
-export { Design, Theme, ThemeRegistry };
+export { Design, mixin, Theme, ThemeRegistry };
